Add support for feature flags, and hide the news section in navigation and home based on this feature flag
