+++ src/services/featureFlagService.ts
@@ -1,0 +1,1 @@
+export const FF_IS_NEWS_ENABLED = false;
